p4-network-monitor/control_plane/controller.py
#!/usr/bin/env python3
"""
P4Runtime Controller for Network Monitoring
Main controller for managing P4 switches and collecting data
"""

import argparse
import grpc
import logging
import sys
import time
from concurrent import futures

# P4Runtime imports (to be installed)
# import p4runtime_sh
# from p4runtime_sh.context import P4RuntimeContext
# from p4runtime_sh.utils import DEFAULT_TIMEOUT

class P4Controller:
    def __init__(self, device_id=0, grpc_port=50051, election_id=(0, 1)):
        """
        Initialize P4Runtime controller
        
        Args:
            device_id: P4 device ID
            grpc_port: gRPC port for P4Runtime
            election_id: Election ID for mastership
        """
        self.device_id = device_id
        self.grpc_port = grpc_port
        self.election_id = election_id
        self.client = None
        
        # Setup logging
        logging.basicConfig(level=logging.INFO)
        self.logger = logging.getLogger(__name__)
    
    def connect(self, switch_address="127.0.0.1"):
        """
        Connect to P4 switch via P4Runtime
        
        Args:
            switch_address: IP address of the P4 switch
        """
        try:
            # TODO: Implement P4Runtime connection
            # self.client = p4runtime_sh.client.P4RuntimeClient(
            #     f'{switch_address}:{self.grpc_port}',
            #     device_id=self.device_id,
            #     election_id=self.election_id
            # )
            self.logger.info(f"Connected to switch at {switch_address}:{self.grpc_port}")
            return True
        except Exception as e:
            self.logger.error(f"Failed to connect to switch: {e}")
            return False
    
    def load_p4_program(self, p4info_file, bmv2_json_file):
        """
        Load P4 program to the switch
        
        Args:
            p4info_file: Path to P4Info file
            bmv2_json_file: Path to BMv2 JSON file
        """
        try:
            # TODO: Implement P4 program loading
            # with open(p4info_file, 'rb') as f:
            #     p4info = f.read()
            # with open(bmv2_json_file, 'rb') as f:
            #     bmv2_json = f.read()
            
            self.logger.info("P4 program loaded successfully")
            return True
        except Exception as e:
            self.logger.error(f"Failed to load P4 program: {e}")
            return False
    
    def install_flow_rules(self):
        """
        Install flow rules for network monitoring
        """
        try:
            # TODO: Implement flow rule installation
            # Example flow rule installation would go here
            self.logger.info("Flow rules installed successfully")
            return True
        except Exception as e:
            self.logger.error(f"Failed to install flow rules: {e}")
            return False
    
    def start_monitoring(self):
        """
        Start monitoring network traffic
        """
        self.logger.info("Starting network monitoring...")
        
        try:
            while True:
                # TODO: Implement monitoring loop
                # - Read counters from switch
                # - Process digest messages
                # - Update flow tables
                # - Export data to database
                
                time.sleep(1)  # Polling interval
                
        except KeyboardInterrupt:
            self.logger.info("Monitoring stopped by user")
        except Exception as e:
            self.logger.error(f"Monitoring error: {e}")
    
    def disconnect(self):
        """
        Disconnect from P4 switch
        """
        if self.client:
            # TODO: Implement proper disconnection
            # self.client.close()
            self.logger.info("Disconnected from switch")

def main():
    parser = argparse.ArgumentParser(description='P4 Network Monitoring Controller')
    parser.add_argument('--switch', default='127.0.0.1', 
                       help='Switch IP address (default: 127.0.0.1)')
    parser.add_argument('--port', type=int, default=50051,
                       help='gRPC port (default: 50051)')
    parser.add_argument('--device-id', type=int, default=0,
                       help='P4 device ID (default: 0)')
    parser.add_argument('--p4info', required=False,
                       help='Path to P4Info file')
    parser.add_argument('--bmv2-json', required=False,
                       help='Path to BMv2 JSON file')
    
    args = parser.parse_args()
    
    # Create and configure controller
    controller = P4Controller(
        device_id=args.device_id,
        grpc_port=args.port
    )
    
    # Connect to switch
    if not controller.connect(args.switch):
        sys.exit(1)
    
    # Load P4 program if provided
    if args.p4info and args.bmv2_json:
        if not controller.load_p4_program(args.p4info, args.bmv2_json):
            sys.exit(1)
    
    # Install flow rules
    if not controller.install_flow_rules():
        sys.exit(1)
    
    try:
        # Start monitoring
        controller.start_monitoring()
    finally:
        # Clean up
        controller.disconnect()

if __name__ == "__main__":
    main()