p4-network-monitor/control_plane/requirements.txt
# P4 and networking
# p4runtime-sh>=1.4.0
grpcio>=1.50.0
grpcio-tools>=1.50.0
# mininet>=2.3.0

# Data processing and storage
influxdb-client>=1.36.0
psycopg2-binary>=2.9.5
redis>=4.5.0
sqlalchemy>=2.0.0
alembic>=1.9.0

# Web framework and APIs
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
websockets>=10.4
pydantic>=1.10.0

# Networking and packet processing
scapy>=2.5.0
netifaces>=0.11.0

# Utilities
pyyaml>=6.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
requests>=2.28.0

# Monitoring and logging
prometheus-client>=0.16.0
structlog>=22.3.0

# Testing and development
pytest>=7.2.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0

# Scientific computing (for analytics)
numpy>=1.24.0
pandas>=1.5.0
scipy>=1.10.0